+++ corehq/ex-submodules/pillow_retry/tests/test_model.py
@@ -1,25 +1,22 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
+
 import sys
-import uuid
 from datetime import datetime
 
 from django.conf import settings
 from django.test import TestCase
-from mock import MagicMock
+from six.moves import range
 
-from pillow_retry.models import PillowError
-from pillow_retry import const
 from pillow_retry.tasks import process_pillow_retry
-from pillowtop import get_all_pillow_configs
+from pillow_retry import const
+from pillow_retry.models import PillowError
 from pillowtop.checkpoints.manager import PillowCheckpoint
-from pillowtop.dao.exceptions import DocumentMissingError
 from pillowtop.feed.couch import change_from_couch_row
 from pillowtop.feed.interface import Change, ChangeMeta
 from pillowtop.feed.mock import RandomChangeFeed
 from pillowtop.processors import PillowProcessor
 from pillowtop.tests.utils import make_fake_constructed_pillow, FakeConstructedPillow
-from six.moves import range
 
 
 def get_ex_tb(message, ex_class=None):
@@ -42,6 +39,7 @@ def GetDocPillow():
         processor=GetDocProcessor(),
     )
 
+
 class GetDocProcessor(PillowProcessor):
     """
     Processor that does absolutely nothing.
@@ -54,7 +52,9 @@ class GetDocProcessor(PillowProcessor):
 
 
 def create_error(change, message='message', attempts=0, pillow=None, ex_class=None):
-    change.metadata = ChangeMeta(data_source_type='couch', data_source_name='test_commcarehq', document_id=change.id)
+    change.metadata = ChangeMeta(
+        data_source_type='couch', data_source_name='test_commcarehq', document_id=change.id
+    )
     error = PillowError.get_or_create(change, pillow or FakePillow())
     for n in range(0, attempts):
         error.add_attempt(*get_ex_tb(message, ex_class=ex_class))
@@ -68,8 +68,8 @@ class PillowRetryTestCase(TestCase):
         self._pillowtops = settings.PILLOWTOPS
         settings.PILLOWTOPS = {
             'tests': [
-                'pillow_retry.tests.FakePillow',
-                'pillow_retry.tests.GetDocPillow',
+                'pillow_retry.tests.test_model.FakePillow',
+                'pillow_retry.tests.test_model.GetDocPillow',
             ]
         }
 
@@ -91,7 +91,7 @@ class PillowRetryTestCase(TestCase):
         self.assertEqual(error.total_attempts, 1)
         self.assertEqual(error.current_attempt, 1)
         self.assertTrue(message in error.error_traceback)
-        self.assertEqual(error.error_type, 'pillow_retry.tests.ExceptionA')
+        self.assertEqual(error.error_type, 'pillow_retry.tests.test_model.ExceptionA')
 
         message = 'ex message2'
         error.add_attempt(*get_ex_tb(message))
@@ -117,7 +117,7 @@ class PillowRetryTestCase(TestCase):
 
     def test_get_errors_to_process(self):
         # Only re-process errors with
-        # current_attempt < setting.PILLOW_RETRY_QUEUE_MAX_PROCESSING_ATTEMPTS
+        # current_attempt < const.PILLOW_RETRY_QUEUE_MAX_PROCESSING_ATTEMPTS
         date = datetime.utcnow()
         for i in range(0, 5):
             error = create_error(_change(id=i), attempts=i+1)
@@ -135,7 +135,6 @@ class PillowRetryTestCase(TestCase):
         self.assertEqual(len(errors), 3)
 
     def test_get_errors_to_process_max_limit(self):
-        # see const.PILLOW_RETRY_MULTI_ATTEMPTS_CUTOFF
         date = datetime.utcnow()
 
         def make_error(id, current_attempt, total_attempts):
@@ -173,7 +172,7 @@ class PillowRetryTestCase(TestCase):
             const.PILLOW_RETRY_MULTI_ATTEMPTS_CUTOFF + 1
         )
 
-        errors = PillowError.get_errors_to_process(date, fetch_full=True).all()
+        errors = PillowError.get_errors_to_process(date).all()
         self.assertEqual(len(errors), 2)
         docs_to_process = {e.doc_id for e in errors}
         self.assertEqual({'to-process1', 'to-process2'}, docs_to_process)
@@ -184,7 +183,7 @@ class PillowRetryTestCase(TestCase):
         error = create_error(change_from_couch_row(change_dict))
         error.save()
         # this used to error out
-        process_pillow_retry(error.id)
+        process_pillow_retry(error)
         with self.assertRaises(PillowError.DoesNotExist):
             PillowError.objects.get(id=error.id)
 
@@ -205,7 +204,7 @@ class PillowRetryTestCase(TestCase):
         for i in range(0, 3):
             error = create_error(_change(id=i), attempts=1)
             if i >= 1:
-                error.total_attempts = PillowError.multi_attempts_cutoff() + 1
+                error.total_attempts = const.PILLOW_RETRY_MULTI_ATTEMPTS_CUTOFF + 1
             error.save()
 
         errors = PillowError.get_errors_to_process(datetime.utcnow()).all()
@@ -224,16 +223,17 @@ class PillowRetryTestCase(TestCase):
             date_last_attempt=datetime.utcnow()
         )
         # make sure this doesn't error
-        process_pillow_retry(error.id)
+        process_pillow_retry(error)
         # and that its total_attempts was bumped above the threshold
-        self.assertTrue(PillowError.objects.get(pk=error.pk).total_attempts > PillowError.multi_attempts_cutoff())
+        error = PillowError.objects.get(pk=error.pk)
+        self.assertTrue(error.total_attempts > const.PILLOW_RETRY_MULTI_ATTEMPTS_CUTOFF)
 
     def test_empty_metadata(self):
         change = _change(id='123')
         error = PillowError.get_or_create(change, GetDocPillow())
         error.save()
 
-        process_pillow_retry(error.id)
+        process_pillow_retry(error)
 
         error = PillowError.objects.get(pk=error.id)
         self.assertEquals(error.total_attempts, 1)
@@ -241,67 +241,6 @@ class PillowRetryTestCase(TestCase):
 
 class ExceptionA(Exception):
     pass
-
-
-class PillowtopRetryAllPillowsTests(TestCase):
-    @classmethod
-    def setUpClass(cls):
-        super(PillowtopRetryAllPillowsTests, cls).setUpClass()
-        cls._PILLOWTOPS = settings.PILLOWTOPS
-        if not settings.PILLOWTOPS:
-            # assumes HqTestSuiteRunner, which blanks this out and saves a copy here
-            settings.PILLOWTOPS = settings._PILLOWTOPS
-
-    @classmethod
-    def tearDownClass(cls):
-        settings.PILLOWTOPS = cls._PILLOWTOPS
-        super(PillowtopRetryAllPillowsTests, cls).tearDownClass()
-
-    def tearDown(self):
-        PillowError.objects.all().delete()
-
-    def test_all_pillows_handle_errors(self):
-        all_pillow_configs = list(get_all_pillow_configs())
-        for pillow_config in all_pillow_configs:
-            self._test_error_logging_for_pillow(pillow_config)
-
-    def _test_error_logging_for_pillow(self, pillow_config):
-        pillow = _pillow_instance_from_config_with_mock_process_change(pillow_config)
-        if pillow.retry_errors:
-            exc_class = Exception
-            exc_class_string = 'exceptions.Exception'
-        else:
-            exc_class = DocumentMissingError
-            exc_class_string = 'pillowtop.dao.exceptions.DocumentMissingError'
-
-        pillow.process_change = MagicMock(side_effect=exc_class(pillow.pillow_id))
-        doc = self._get_random_doc()
-        pillow.process_with_error_handling(Change(id=doc['id'], sequence_id='3', document=doc))
-
-        errors = PillowError.objects.filter(pillow=pillow.pillow_id).all()
-        self.assertEqual(1, len(errors), pillow_config)
-        error = errors[0]
-        self.assertEqual(error.doc_id, doc['id'], pillow_config)
-        self.assertEqual(exc_class_string, error.error_type)
-        self.assertIn(pillow.pillow_id, error.error_traceback)
-
-    def _get_random_doc(self):
-        return {
-            'id': uuid.uuid4().hex,
-            'doc_type': 'CommCareCase',
-            'type': 'mother',
-            'domain': 'pillow-retry-domain',
-        }
-
-
-def _pillow_instance_from_config_with_mock_process_change(pillow_config):
-    pillow_class = pillow_config.get_class()
-    if pillow_config.instance_generator is None:
-        instance = pillow_class()
-    else:
-        instance = pillow_config.get_instance()
-
-    return instance
 
 
 def _change(id):
